# Thank you for your contribution to the Chem4Word project.

## Please enter description of what these changes are for
Pushers now drop nascent bonds on shift-drag
**Describe the changes being made to the code and why**

## After submitting this PR for code review, please make sure:
- [x] "Describe the changes being made to the code and why" above has been replaced
- [x] All new files have the standard license headers
- [x] CodeMaid has been run against all new or updated files
- [x] The code builds clean (please check Debug and Setup configurations) without any errors or new warnings
- [x] All NuGet packages have been updated and /README.md updated
- [x] All Unit Tests pass
- [x] You have followed the [Chem4Word Coding Standards](https://chem4word.visualstudio.com/C4W-VNext/_wiki/wikis/C4W-VNext.wiki/2/Coding-Standards)
- [x] You have updated the change history (/src/Chem4Word.V3/Data/Chem4Word-Versions.xml)
- [ ] If appropriate you have updated the manual (/docs/Chem4Word-Version3-3-User-Manual.docx)
- [x] The build pipeline is green


Related work items: #1279

+++ src/Common/Chem4Word.Core/SqLite/Library.cs
@@ -54,8 +54,6 @@ namespace Chem4Word.Core.SqLite
                 // CoPilot suggested fix for occasional error, with tweak to ensure path only set once.
                 //  "Unable to find an entry point named 'SI3eae3b91c35710f2' in DLL 'SQLite.Interop.dll'."
 
-                // ToDo: Add some telemetry ...
-
                 string executingPath = Path.GetDirectoryName(Assembly.GetExecutingAssembly().Location);
                 bool is64 = Environment.Is64BitProcess;
                 string source = Path.Combine(AppDomain.CurrentDomain.BaseDirectory, is64 ? "x64" : "x86", "SQLite.Interop.dll");
@@ -63,7 +61,7 @@ namespace Chem4Word.Core.SqLite
                 {
                     string dest = Path.Combine(executingPath, "SQLite.Interop.dll");
 
-                    if (!File.Exists(dest))
+                    if (File.Exists(source) && !File.Exists(dest))
                     {
                         File.Copy(source, dest, overwrite: true);
                     }
